refactor : domain folder 분리
domain folder 생성 후 domain들을 하위 풀더로 이동
