+++ server/src/domain/auth/auth.router.ts
@@ -1,12 +1,12 @@
 import { ACCESS_DENIED } from 'common/error';
 import Router from 'koa-router';
 import { Context } from 'koa';
-import userRepository from '@/user/user.repository';
+import userRepository from '@/domain/user/user.repository';
 import { v4 } from 'uuid';
-import { JwtConfig } from '@config/index';
-import UserService from '../user/user.service';
-import jwtAuthorize from '../middleware/jwt-authorize';
-import OAuthClient from './oauth-client';
+import { JwtConfig } from '@/config/index';
+import UserService from '@/domain/user/user.service';
+import jwtAuthorize from '@/middleware/jwt-authorize';
+import OAuthClient from '@/lib/oauth-client';
 import JwtUtils from './utils/jwt-utils';
 
 class AuthRouter extends Router {
