+++ server/src/domain/auth/types/decoded-jwt.ts
