+++ server/src/domain/auth/types/kakao-user-dto.ts
