+++ server/src/domain/auth/types/naver-user-dto.ts
