+++ server/src/domain/auth/types/social-user-dto.ts
