+++ server/src/domain/auth/types/user-dto.ts
