+++ server/src/domain/category/.gitkeep
@@ -1,0 +1,1 @@
+category domain folder
