+++ server/src/domain/fixed-expenditure/fixed-expenditure.repository.ts
