+++ server/src/domain/fixed-expenditure/types.ts
