+++ server/src/domain/payment/.gitkeep
@@ -1,0 +1,1 @@
+payment domain folder
