+++ server/src/domain/transaction/transaction.repository.ts
@@ -1,5 +1,5 @@
 import { getRepository, Repository } from 'typeorm';
-import TransactionEntity from '@entity/transaction.entity';
+import TransactionEntity from '@/entity/transaction.entity';
 
 const getTransactionRepository = (): Repository<TransactionEntity> => {
   return getRepository(TransactionEntity);
