+++ server/src/domain/user/types/user.ts
