+++ server/src/middleware/jwt-authorize.ts
@@ -1,10 +1,10 @@
 import { Next } from 'koa';
 import { Context } from 'vm';
 import { getRepository } from 'typeorm';
-import JwtUtils from '@auth/utils/jwt-utils';
+import JwtUtils from '@/domain/auth/utils/jwt-utils';
 import { JwtConfig } from '@config/index';
 import UserEntity from '../entity/user.entity';
-import UserDTO from '../auth/types/user-dto';
+import UserDTO from '../domain/auth/types/user-dto';
 
 const jwtAuthorize = async (ctx: Context, next: Next): Promise<void> => {
   const token = ctx.cookies.get('jwt');
